Update bootstrap-modal.js
To run the script in my need, I needed to use the parameter "manager" and order the popup's smug written inside the corresponding form. So, to open a popup with form inside another form, a conflict occurred for closing them, which was corrected with my contribution. I am studying how to solve this also for the ESC. The default behavior is not changed by my changes.